refactor(inspection) Removed class.
