+++ blackduck-artifactory-common/src/test/groovy/com/synopsys/integration/blackduck/artifactory/legacy/metadata/VersionBomComponentLinkTest.java
@@ -29,7 +29,6 @@ import org.junit.Test;
 import com.synopsys.integration.blackduck.api.UriSingleResponse;
 import com.synopsys.integration.blackduck.api.generated.view.ComponentVersionView;
 import com.synopsys.integration.blackduck.api.generated.view.ProjectVersionView;
-import com.synopsys.integration.blackduck.artifactory.modules.inspection.notifications.CompositeComponentManager;
 
 public class VersionBomComponentLinkTest {
     @Test
